+++ semestr3/ZMPO/ZMPO_4/ZMPO_4/Creature.h
@@ -5,7 +5,7 @@
 
 class Creature {
 public:
-	Creature(int genotypeLength);
+	Creature(size_t genotypeLength);
 	void SetFitness(double fitness);
 	double GetFitness();
 	Creature* CrossWith(const Creature* other);
